ui,nodes: better lifetime management
When a node connects to the GUI (server) for the first time, it's
identified by the connection (src_port:address:dst_port).
We save it to identify and manage the node.

When a node disconnects from the server (GUI), the session and the
node address are the same, so we update the status accordingly.

However, if a node does not respond for some time (minutes, hours,...), and then
reconnects again, it'll connect with a new sesion address
(src_port:addr:dst_port).
In this situation, the old session still exists on the server, and when
it exceeds the maximum lifetime, the server closes it and reporting the old
session address.

Therefore, to prevent the current node from being marked as 'offline',
we ignore this notification.